# sstring: use char_traits<char>::compare() when possible