C++ standard library specializes char_traits<char_type>
for all builtin char types. see
http://eel.is/c++draft/char.traits.specializations.general

and it's found that both libstdc++ and libc++ uses memcmp()
and wmemcmp to implement the specialization of
char_traits<char_type>::compare(), where char_type is char,
wchar_t and char8_t for better performance.

but in our typical use cases of basic_sstring, in addition to
using seastar::sstring, we also use seastar::basic_sstring<int8_t,..>
to present a string of bytes. but if int8_t is not an alias of
char or char8_t, the performance of comparison degrades
dramatically, as libstdc++ and libc++ would fall back to the
loop-based implementation in this case.

so, in this change, instead of calling char_traits<char_type>::compare()
directly, we check if the char_type's signness is identical to that
of `char`, and cast to the matched one for better performance.

before/after this change:

| test | iterations |median | mad | min | max | allocs | tasks | inst |
| --- | --- | --- | --- | --- | --- | --- | --- | --- |
| random_strings.compare | 67727389 | 14.406ns | 0.110ns | 13.301ns | 14.671ns | 0.000 | 0.000 | 266.0 |
| random_strings.compare | 436572368 | 2.216ns | 0.013ns | 2.199ns | 2.239ns | 0.000 | 0.000 | 54.0 |

with clang-17, CMAKE_BUILD_TYPE=RelWithDebInfo.