@tchaikov 2023-06-05
inspired by https://github.com/redpanda-data/redpanda/pull/10385