@tchaikov 2023-06-09
changelog

- instead of comparing the size of `char` and `char_type`, compare their signness of them

the perf test result is identical. as in Clang and GCC, char is signed by default.

@michoecho hi Michał, could you take another look?