@michoecho 2023-06-09
Just an experimental proof to confirm my point:

https://godbolt.org/z/fqG5PWs4E

```
#include <string>
#include <cassert>

int main() {
    // char is signed on x86.
    static_assert(std::is_signed_v<signed char> == std::is_signed_v<char>);
    // `signed char` strings have signed comparison.
    static_assert(std::basic_string<signed char>({-1}) < std::basic_string<signed char>({0}));
    // `char` is signed on this platform, but `char` strings have unsigned comparison anyway. So the below assert fails.
    static_assert(std::basic_string<char>({-1}) < std::basic_string<char>({0}));
}
```